requirements.txt
numpy
matplotlib
scikit-learn
fire
lightning
torchinfo
wandb>=0.12.10
jiwer
biopython
tqdm
pillow
pyyaml
beautifulsoup4
